Symfony setup using a custom Bundle
